Include the driver/helpers directory
Just add a README file for this, to keep the directory present.
This will cause a warning, as the Uploader does not know what to do with the README file, but at least you can start developing right away.

Another possibility would be to add a post-install script that creates this directory automagically...